# .md/API-CONTRACT.yaml
openapi: 3.0.3
info:
  title: Turma do Rola - Comary — Contrato de API
  version: 0.12.0
  description: >-
    Contrato publicado incrementalmente, endpoint a endpoint, assim que cada
    contrato fica estável (TASK.md Seção 1.7 / GUARDRAILS.md regra implícita
    de rastreabilidade de API) — não é preciso esperar a tarefa de Backend
    inteira terminar. Frontend/Mobile começam a consumir assim que um
    endpoint aparece aqui, mesmo que a implementação completa da tarefa
    ainda esteja em andamento. Mudar um contrato já publicado é permitido,
    mas nunca silenciosamente — toda mudança de contrato incompatível
    incrementa `info.version` e é registrada na seção "Changelog" no final
    deste arquivo.

    Duas famílias de endpoint neste sistema:
      1. **PostgREST autogerado pelo Supabase** sobre as views públicas
         curadas (`ranking_publico`, `presenca_mensal_publica`) — consumido
         **diretamente** pelo Frontend com a chave `anon` do Supabase, sem
         passar pela camada de API própria (SDD.md Seção 7.5, ADR-005). É o
         único caso em que o Frontend fala com o Supabase sem intermediação
         de um Route Handler — decisão já tomada no SDD.md, não uma exceção
         nova. Documentado aqui (paths desta seção) porque também é
         superfície pública real, exposta à internet.
      2. **Route Handlers do Next.js (App Router)**, publicados a partir de
         BE-04 em diante — toda escrita e toda leitura autenticada da área
...
servers:
...
components:
  securitySchemes:
    supabaseAnonKey:
      type: apiKey
      in: header
      name: apikey
      description: >-
        Chave pública (`anon`) do Supabase — único credential usado pelas
        páginas públicas (T02/T03). Nunca a `service role` (TASK.md Seção
        1.2/1.7, GUARDRAILS.md regra 7).
    sessionCookie:
      type: apiKey
      in: cookie
      name: sessao_interna
      description: >-
        Cookie de sessão emitido por `POST /api/auth/login` (`httpOnly`,
        `Secure`, `SameSite=Strict`, TTL 8-12h — ADR-004, BE-04). Único
        mecanismo de autenticação da área interna (modelo binário, sem
        RBAC — SDD.md Seção 7.2): sessão válida = acesso total. Toda rota
        de escrita futura desta área (BE-06 em diante) deve declarar este
        esquema de segurança.
  responses:
    SessaoInvalida:
      description: >-
        Sem sessão válida (`middleware.ts`, BE-04, GUARDRAILS.md regra 17) —
        toda rota de escrita da área interna, exceto `/api/auth/login` e
        `/api/auth/logout`, retorna isto quando o cookie `sessao_interna`
        está ausente, expirado ou com assinatura inválida. Resposta
        reutilizável — toda rota de escrita futura (BE-06 em diante) deve
        referenciar `#/components/responses/SessaoInvalida` para o caso
        401, em vez de redefinir o formato.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErroSessaoInvalida"
  schemas:
...
    ErroSessaoInvalida:
      type: object
      description: >-
        Corpo de erro `401` emitido pelo middleware de sessão
        (`middleware.ts`, BE-04) para qualquer rota de escrita sem sessão
        válida — formato distinto do erro de login (contexto diferente:
        aqui é "sessão válida ou não", não "senha certa ou errada").
      properties:
        error:
          type: string
          example: "Sessão inválida ou expirada."
      required: [error]
      additionalProperties: false
    AtletaBody:
      type: object
      description: >-
        Corpo de `POST /api/atletas` e `PUT /api/atletas/{id}` (BE-06).
        Campos em `snake_case`, mesma convenção já usada pelas respostas de
        `RankingPublicoItem`/`PresencaMensalPublicaItem` no restante deste
        contrato (decisão de consistência, TASK.md Seção 1.0 — documentada
        na nota de status de BE-06 do `TASK.md`, não escalada).
      properties:
        nome_completo:
          type: string
          minLength: 1
        apelido_exibicao:
          type: string
          minLength: 1
          description: >-
            Opcional — quando ausente/em branco, o Backend deriva o primeiro
            "token" de `nome_completo` (RF-01.2/RN-06).
        contato:
          type: string
          minLength: 1
          nullable: true
        data_nascimento:
          type: string
          format: date
          description: >-
            Obrigatório neste formulário (RF-01.1, UX-SPEC.md T04 marca com
            `*`) — não pode ser no futuro. Coluna correspondente em
            `app.atleta` é `nullable` no banco só para acomodar registro
            migrado do legado sem essa informação (RF-08.3/BE-15), não para
            tornar o campo opcional aqui.
        consentimento_responsavel_obtido:
          type: boolean
          default: false
          description: >-
            Obrigatoriamente `true` quando `data_nascimento` resultar em
            idade < 18 anos na data da requisição — caso contrário a API
            recusa com `400` (RF-01.3/RN-02).
        pontuacao_inicial:
          type: number
          minimum: 0
        confirmar_duplicidade:
          type: boolean
          default: false
          description: >-
            Quando `nome_completo` coincide (normalizado — trim/espaço/caixa)
            com o de outro atleta ativo e este campo não vier `true`, a API
            recusa com `409` (alerta de duplicidade, RF-01.5) em vez de
            gravar. O Frontend reenvia a mesma requisição com
            `confirmar_duplicidade: true` depois que o organizador confirmar
            no modal (UX-SPEC.md T04) — nunca um bloqueio definitivo, não há
            `UNIQUE` em `nome_completo` no schema (BE-02).
      required: [nome_completo, data_nascimento, pontuacao_inicial]
      additionalProperties: false
    AtletaResponse:
      type: object
      description: >-
        Resposta de `GET/POST /api/atletas` (lista/criação) e
        `GET/PUT /api/atletas/{id}` (BE-06). Nunca devolvida a cliente sem
        sessão válida (RN-01/GUARDRAILS.md regra 19, ver tag `interno`).
      properties:
        id:
          type: string
          format: uuid
        nome_completo:
          type: string
        apelido_exibicao:
          type: string
        contato:
          type: string
          nullable: true
        data_nascimento:
          type: string
          format: date
          nullable: true
        consentimento_responsavel_obtido:
          type: boolean
        pontuacao_inicial:
          type: number
        ativo:
          type: boolean
        anonimizado_em:
          type: string
          format: date-time
          nullable: true
          description: >-
            Preenchido pela função `anonimizar_atleta` (ADR-011, BE-07) —
            ver `POST /api/atletas/{id}/anonimizar`.
        criado_em:
          type: string
          format: date-time
        nivel_tecnico:
          type: number
          description: >-
            RN-03 — média de pontos de eventos de jogo (gol/cartão) por
            rodada com presença (`app.atleta_nivel_tecnico`, migration
            BE-06). Fallback = `pontuacao_inicial` para atleta sem nenhuma
            presença registrada. Nunca um campo de entrada do formulário
            (RF-01.4 — sempre derivado).
        rodadas_presentes:
          type: integer
          description: Denominador usado no cálculo de `nivel_tecnico` acima.
      required:
        [
          id,
          nome_completo,
          apelido_exibicao,
          contato,
          data_nascimento,
          consentimento_responsavel_obtido,
          pontuacao_inicial,
          ativo,
          anonimizado_em,
          criado_em,
          nivel_tecnico,
          rodadas_presentes,
        ]
      additionalProperties: false
    ErroValidacaoAtleta:
      type: object
      description: >-
        Corpo de erro `400` de `POST/PUT /api/atletas*` — corpo malformado ou
        que falha `AtletaBody` (inclui o bloqueio de RF-01.3: menor de 18 sem
        `consentimento_responsavel_obtido`, reportado como um item de
        `detalhes` com `path: ["consentimento_responsavel_obtido"]`).
      properties:
        error:
          type: string
          example: "Requisição inválida."
        detalhes:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
            required: [path, message]
      required: [error]
      additionalProperties: false
    ErroDuplicidadeAtleta:
      type: object
      description: >-
        Corpo de erro `409` de `POST/PUT /api/atletas*` (RF-01.5) — emitido
        quando `nome_completo` (normalizado) coincide com o de outro atleta
        ativo e a requisição não trouxe `confirmar_duplicidade: true`.
      properties:
        error:
          type: string
          example: "duplicidade"
        atletas_duplicados:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              nome_completo:
                type: string
            required: [id, nome_completo]
      required: [error, atletas_duplicados]
      additionalProperties: false
...
    RestricaoBody:
      type: object
      description: >-
        Corpo de `POST /api/restricoes` e `PUT /api/restricoes/{id}` (BE-12,
        RF-05.5). Sem campo de "motivo" (`UX-SPEC.md` T10 — "sem campo de
        motivo obrigatório no PRD-TECNICO... fora de escopo desta release").
      properties:
        atleta_a_id:
          type: string
          format: uuid
        atleta_b_id:
          type: string
          format: uuid
          description: >-
            Deve ser diferente de `atleta_a_id` — `400` caso contrário
            (`restricao_obrigatoria_atletas_distintos_check` no banco,
            BE-02, replicado aqui como validação de API).
      required: [atleta_a_id, atleta_b_id]
      additionalProperties: false
    RestricaoObrigatoriaResponse:
      type: object
      description: >-
        Resposta de todos os endpoints de `/api/restricoes*` (BE-12).
        `atleta_a_nome`/`atleta_b_nome` (via `apelido_exibicao`, RN-06)
        espelham de propósito os mesmos nomes de campo do contrato
        `restricoes_conflitantes` do ADR-010 — `BE-11` consome
        `app.restricao_obrigatoria` para montar aquele contrato a partir da
        mesma tabela.
      properties:
        id:
          type: string
          format: uuid
        atleta_a_id:
          type: string
          format: uuid
        atleta_a_nome:
          type: string
          description: "`apelido_exibicao` (RN-06) do atleta em `atleta_a_id`."
        atleta_b_id:
          type: string
          format: uuid
        atleta_b_nome:
          type: string
          description: "`apelido_exibicao` (RN-06) do atleta em `atleta_b_id`."
        ativo:
          type: boolean
        desativado_em:
          type: string
          format: date-time
          nullable: true
          description: >-
            Preenchido na primeira desativação (RN-11) — nunca sobrescrito
            por chamadas repetidas de `POST .../desativar` (idempotência,
            preserva o registro histórico real); volta a `null` depois de
            `POST .../reativar`.
        criado_em:
          type: string
          format: date-time
      required:
        [
          id,
          atleta_a_id,
          atleta_a_nome,
          atleta_b_id,
          atleta_b_nome,
          ativo,
          desativado_em,
          criado_em,
        ]
      additionalProperties: false
    ErroValidacaoRestricao:
      type: object
      description: >-
        Corpo de erro `400` de `POST/PUT /api/restricoes*` — corpo
        malformado, `uuid` inválido, ou `atleta_a_id === atleta_b_id`.
      properties:
        error:
          type: string
          example: "Requisição inválida."
        detalhes:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
            required: [path, message]
      required: [error]
      additionalProperties: false
    ErroAtletaReferenciadoNaoEncontrado:
      type: object
      description: >-
        Corpo de erro `404` de `POST/PUT /api/restricoes*` quando
        `atleta_a_id`/`atleta_b_id` não corresponde a nenhum `app.atleta`
        existente — devolvido pela API em vez de deixar a `FK` do banco
        (`references app.atleta (id)`, BE-02) estourar como erro genérico
        `23503`.
      properties:
        error:
          type: string
          example: "Atleta não encontrado."
        atleta_id:
          type: string
          format: uuid
          description: id do atleta que não foi encontrado (o primeiro dos dois, se ambos faltarem).
      required: [error, atleta_id]
      additionalProperties: false
...
    SugestaoTimesBody:
      type: object
      description: >-
        Corpo de `POST /api/times/sugestao` (BE-11, RF-05.1/RF-05.3). `N`
        (`quantidade_times`) é parametrizável de verdade no backend (ADR-010/
        TASK.md Seção 1.4) mesmo a interface desta release só enviando `2`
        (TASK.md Seção 6.2 item 1). Não referencia `rodada_id` — BE-11 não
        depende de BE-08 (TASK.md Seção 4.1) — a lista de presentes é
        informada diretamente aqui, não derivada de uma rodada já lançada.
      properties:
        atletas_ids:
          type: array
          minItems: 2
          items:
            type: string
            format: uuid
          description: Presentes desta montagem (RF-05.1) — sem ids repetidos.
        quantidade_times:
          type: integer
          minimum: 2
          maximum: 10
          description: >-
            `N` — quantidade de times a gerar. Não pode ser maior que o
            tamanho de `atletas_ids`. Teto de 10 é decisão de detalhe (não
            escalada, sem limite literal em RF-05.1/ADR-010).
      required: [atletas_ids, quantidade_times]
      additionalProperties: false
    AtletaMontadoResponse:
      type: object
      description: Um atleta dentro de um time sugerido (`SugestaoTimesOkResponse`).
      properties:
        atleta_id:
          type: string
          format: uuid
        apelido_exibicao:
          type: string
        nivel_tecnico:
          type: number
          description: RN-03, via `app.atleta_nivel_tecnico` (mesma view de BE-06).
        idade:
          type: integer
          nullable: true
          description: >-
            `null` quando o atleta não tem `data_nascimento` cadastrado
            (registro migrado do legado, RF-08.3) — excluído do cálculo de
            `idade_media` do time, nunca tratado como `0`.
      required: [atleta_id, apelido_exibicao, nivel_tecnico, idade]
      additionalProperties: false
    TimeMontadoResponse:
      type: object
      description: Um time sugerido, dentro de `SugestaoTimesOkResponse.times`.
      properties:
        indice:
          type: integer
          minimum: 0
          description: 0-based — sem significado próprio além do rótulo ("Time A" = índice 0 no Frontend).
        atletas:
          type: array
          items:
            $ref: "#/components/schemas/AtletaMontadoResponse"
        nivel_tecnico_medio:
          type: number
          description: Média de `nivel_tecnico` dos atletas do time (RF-05.3), arredondada a 2 casas.
        idade_media:
          type: number
          nullable: true
          description: >-
            Média de idade dos atletas do time com `data_nascimento`
            conhecido (RF-05.3), arredondada a 1 casa — `null` se nenhum
            atleta do time tiver `data_nascimento`.
      required: [indice, atletas, nivel_tecnico_medio, idade_media]
      additionalProperties: false
    SugestaoTimesOkResponse:
      type: object
      description: >-
        Resposta de sucesso de `POST /api/times/sugestao` — a divisão
        respeita 100% das restrições obrigatórias ativas entre os
        presentes (RF-05.1); nível técnico + idade usados como soft
        constraint (RF-05.3, ADR-007 Fase 2 — busca local, "boa o
        suficiente", não necessariamente ótima global). É sempre uma
        SUGESTÃO (RF-05.4 — o organizador pode ajustar manualmente antes de
        confirmar); esta API não persiste nada em `app.time`/`app.time_atleta`.
      properties:
        status:
          type: string
          enum: [ok]
        quantidade_times_solicitada:
          type: integer
        times:
          type: array
          items:
            $ref: "#/components/schemas/TimeMontadoResponse"
      required: [status, quantidade_times_solicitada, times]
      additionalProperties: false
    RestricaoConflitanteItem:
      type: object
      description: >-
        Um par em conflito, contrato EXATO do ADR-010 — um item por aresta
        do componente conexo que falhou a coloração em `N` times (nenhum
        formato simplificado, ADR-010 Opção B foi rejeitada).
      properties:
        restricao_id:
          type: string
          format: uuid
        atleta_a_id:
          type: string
          format: uuid
        atleta_a_nome:
          type: string
          description: "`apelido_exibicao` (RN-06) — mesmo campo já publicado em `RestricaoObrigatoriaResponse` (BE-12)."
        atleta_b_id:
          type: string
          format: uuid
        atleta_b_nome:
          type: string
        motivo:
          type: string
          enum: [restricao_obrigatoria_ativa]
        grupo_conflito:
          type: integer
          minimum: 1
          description: 1-based — agrupa pares do mesmo componente conexo (ADR-010).
      required:
        [
          restricao_id,
          atleta_a_id,
          atleta_a_nome,
          atleta_b_id,
          atleta_b_nome,
          motivo,
          grupo_conflito,
        ]
      additionalProperties: false
    GrupoConflitoItem:
      type: object
      description: Um componente conexo sem coloração válida em `N` times (ADR-010).
      properties:
        grupo_conflito:
          type: integer
          minimum: 1
        atletas_ids:
          type: array
          items:
            type: string
            format: uuid
        quantidade_times_solicitada:
          type: integer
        mensagem:
          type: string
          example: >-
            Com 2 time(s) disponível(is), não é possível separar os 3
            atletas deste grupo sem que alguma restrição obrigatória fique
            violada.
      required: [grupo_conflito, atletas_ids, quantidade_times_solicitada, mensagem]
      additionalProperties: false
    SugestaoTimesConflitoResponse:
      type: object
      description: >-
        Resposta de `POST /api/times/sugestao` quando nenhuma divisão em `N`
        times satisfaz 100% das restrições obrigatórias ativas (RF-05.2) —
        contrato EXATO do ADR-010 (`restricoes_conflitantes`/
        `grupos_conflito`), decomposto por componente conexo do grafo de
        restrições (union-find). Um resultado VÁLIDO do algoritmo, não um
        erro técnico — por isso HTTP `200`, igual ao caso `status: "ok"`
        (RF-05.2: "informar, não travar").
      properties:
        status:
          type: string
          enum: [conflito]
        restricoes_conflitantes:
          type: array
          items:
            $ref: "#/components/schemas/RestricaoConflitanteItem"
        grupos_conflito:
          type: array
          items:
            $ref: "#/components/schemas/GrupoConflitoItem"
      required: [status, restricoes_conflitantes, grupos_conflito]
      additionalProperties: false
    ErroValidacaoSugestaoTimes:
      type: object
      description: >-
        Corpo de erro `400` de `POST /api/times/sugestao` — corpo
        malformado, `uuid` inválido, `atletas_ids` com menos de 2 itens ou
        com id repetido, `quantidade_times` fora de `[2, 10]` ou maior que
        `atletas_ids.length`.
      properties:
        error:
          type: string
          example: "Requisição inválida."
        detalhes:
          type: array
          items:
            type: object
            properties:
              path:
                type: array
                items: {}
              message:
                type: string
            required: [path, message]
      required: [error]
      additionalProperties: false
    ErroFalhaTecnicaMontagemTimes:
      type: object
      description: >-
        Corpo de erro `500` de `POST /api/times/sugestao` — a execução do
        algoritmo (backtracking, ADR-007) excedeu o orçamento de tempo
        configurado (`TIMEOUT_MONTAGEM_MS`, 8s — TASK.md Seção 6.2 item 3,
        margem de segurança abaixo do limite prático de execução serverless
        do tier gratuito/hobby da Vercel). Nunca um resultado parcial/
        incompleto — nem `"ok"` nem `"conflito"` — porque a Fase 1 não teve
        tempo de provar nenhum dos dois. Reaproveitado pelo estado genérico
        de "falha técnica real" já previsto pelo `UX-SPEC.md` para T09
        (Seção 4), sem tela nova.
      properties:
        error:
          type: string
          enum: [falha_tecnica]
        message:
          type: string
      required: [error, message]
      additionalProperties: false
...
paths:
...
  /api/atletas:
    get:
      tags: [interno]
      summary: Lista atletas (BE-06, RF-01)
      description: >-
        Exige sessão válida mesmo sendo leitura (ver tag `interno`) —
        devolve `contato`/`data_nascimento`. Cada item inclui o nível
        técnico (RN-03) já calculado (`nivel_tecnico`/`rodadas_presentes`).
        Sem paginação nesta versão (porte do grupo — dezenas de atletas, não
        milhares, TASK.md Seção 1.0).
      operationId: getAtletas
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Lista de atletas, ordenada por `nome_completo` asc.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AtletaResponse"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
    post:
      tags: [interno]
      summary: Cadastra um atleta (BE-06, RF-01.1)
      description: >-
        Inicializa o saldo acumulado do atleta com `pontuacao_inicial`
        (RF-01.1/RN-10). Bloqueia com `400` se `data_nascimento` resultar em
        idade < 18 anos e `consentimento_responsavel_obtido` não vier `true`
        (RF-01.3/RN-02). Alerta de duplicidade de `nome_completo` (RF-01.5)
        — ver `409` abaixo.
      operationId: postAtletas
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtletaBody"
      responses:
        "201":
          description: Atleta criado. `nivel_tecnico` = `pontuacao_inicial` (nenhuma presença ainda, fallback RN-03).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtletaResponse"
        "400":
          description: >-
            Corpo malformado/inválido — inclui o bloqueio de RF-01.3 (menor
            de 18 sem consentimento marcado) como um item de `detalhes` com
            `path: ["consentimento_responsavel_obtido"]`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoAtleta"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "409":
          description: >-
            `nome_completo` (normalizado) coincide com o de outro atleta
            ativo e a requisição não trouxe `confirmar_duplicidade: true`
            (RF-01.5) — nenhum registro foi criado. Reenviar a mesma
            requisição com `confirmar_duplicidade: true` para confirmar e
            criar mesmo assim.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroDuplicidadeAtleta"
...
  /api/restricoes:
    get:
      tags: [interno]
      summary: Lista restrições obrigatórias, ativas e desativadas (BE-12, RF-05.5)
      description: >-
        Exige sessão válida mesmo sendo leitura (mesmo racional de
        `GET /api/log-auditoria` — feature exclusiva da área interna, T10 do
        `UX-SPEC.md`; `middleware.ts`, `INTERNAL_READ_PROTECTED_PREFIXES`).
        Devolve TODAS as restrições, ativas e desativadas (RN-11 — histórico
        nunca some da lista, apenas `ativo`/`desativado_em` mudam),
        ordenadas com as ativas primeiro, mais recente para mais antiga
        dentro de cada grupo (decisão de detalhe, não exigida literalmente
        pelo critério de aceite). Cada item já vem com
        `atleta_a_nome`/`atleta_b_nome` (RN-06) resolvidos, evitando
        chamadas extras no Frontend (T10).
      operationId: getRestricoes
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Lista de restrições.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestricaoObrigatoriaResponse"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
    post:
      tags: [interno]
      summary: Cadastra uma restrição obrigatória entre dois atletas (BE-12, RF-05.5)
      description: >-
        Cria com `ativo=true`/`desativado_em=null`. Recusa com `400` se
        `atleta_a_id === atleta_b_id`; com `404` se algum dos dois ids não
        corresponder a um atleta existente. Sem verificação de duplicidade
        de par já cadastrado (decisão de detalhe, TASK.md Seção 1.0 — não
        exigida por RF-05.5/RN-11, diferente de RF-01.5 para
        `nome_completo` de atleta; um par duplicado é inofensivo para o
        grafo de conflito do ADR-010, aresta repetida não muda o resultado
        da checagem por componente conexo). RN-12: qualquer sessão válida
        pode chamar este endpoint — não há hierarquia nem campo de autor
        individual em nenhum lugar deste contrato.
      operationId: postRestricoes
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestricaoBody"
      responses:
        "201":
          description: Restrição criada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestricaoObrigatoriaResponse"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoRestricao"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "404":
          description: "`atleta_a_id` ou `atleta_b_id` não corresponde a nenhum atleta existente."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroAtletaReferenciadoNaoEncontrado"
...
  /api/times/sugestao:
    post:
      tags: [interno]
      summary: >-
        Gera uma sugestão de divisão de times para os presentes informados
        (BE-11, RF-05.1 a RF-05.3, ADR-007/ADR-010)
      description: >-
        Heurística determinística de duas fases (ADR-007): (1) backtracking
        com poda, por componente conexo do grafo de restrições obrigatórias
        ativas (union-find, ADR-010), garantindo 100% das restrições
        obrigatórias entre os presentes OU provando que nenhuma divisão em
        `N` times as satisfaz simultaneamente (RF-05.1/RF-05.2); (2) busca
        local (swap iterativo entre times) minimizando a diferença agregada
        de nível técnico (RN-03) + idade entre os times (RF-05.3,
        "melhor esforço", não necessariamente ótimo global). Não persiste
        nada em `app.time`/`app.time_atleta` — devolve só a SUGESTÃO
        (RF-05.4, "o organizador pode ajustar manualmente antes de
        confirmar"); persistir a divisão confirmada é um gap desta
        decomposição, sinalizado na nota de status de `BE-11` no `TASK.md`,
        não coberto por este endpoint. Guard de timeout de 8s (TASK.md
        Seção 6.2 item 3) — ver `500` abaixo.
      operationId: postSugestaoTimes
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SugestaoTimesBody"
      responses:
        "200":
          description: >-
            Geração concluída dentro do orçamento de tempo — `status: "ok"`
            (divisão válida, ver `SugestaoTimesOkResponse`) ou `status:
            "conflito"` (nenhuma divisão em `N` times satisfaz 100% das
            restrições obrigatórias, ver `SugestaoTimesConflitoResponse`).
            Os dois são resultados VÁLIDOS do algoritmo (RF-05.2) — nunca um
            erro HTTP.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/SugestaoTimesOkResponse"
                  - $ref: "#/components/schemas/SugestaoTimesConflitoResponse"
                discriminator:
                  propertyName: status
                  mapping:
                    ok: "#/components/schemas/SugestaoTimesOkResponse"
                    conflito: "#/components/schemas/SugestaoTimesConflitoResponse"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroValidacaoSugestaoTimes"
        "401":
          $ref: "#/components/responses/SessaoInvalida"
        "404":
          description: >-
            Algum item de `atletas_ids` não corresponde a nenhum atleta
            existente — mesmo contrato de erro de `POST/PUT /api/restricoes*`
            (BE-12), reaproveitado aqui por ser a mesma situação (id
            referenciado em corpo de requisição que não existe em
            `app.atleta`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroAtletaReferenciadoNaoEncontrado"
        "500":
          description: >-
            Execução acima do timeout configurado (TASK.md Seção 6.2 item 3)
            — falha técnica real, nunca a função serverless travada sem
            resposta.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErroFalhaTecnicaMontagemTimes"
...